+++ rascal2/widgets/project/slider_view.py
